moved loader to its own source module
